Fix Nemotron-30B garbage output: use layer.activation instead of hardcoded F.silu
Root cause: NemotronH MoE uses relu2_no_mul (squared ReLU) activation, but
GLQFusedMoEMethod.apply() hardcoded F.silu for all models. This corrupted
every expert output, producing nonsensical text.

- Add _apply_activation() dispatcher supporting silu/gelu/relu2 (gated + no_mul)
- Read activation from layer.activation (str in vLLM 0.16, enum in 0.18+)
- Fix bf16/fp16 dtype mismatch in dense Mamba layer apply path

+++ glq_vllm/linear_method.py
@@ -540,7 +540,7 @@ class GLQLinearMethod(LinearMethodBase):
     ) -> torch.Tensor:
         # Mamba layers dequantized to dense — standard matmul
         if getattr(layer, '_glq_use_dense', False):
-            return F.linear(x, layer.weight, bias)
+            return F.linear(x, layer.weight.to(x.dtype), bias)
         orig_shape = x.shape
         in_features = layer.glq_in_features
         x = x.reshape(-1, in_features)
